+++ src/main/java/com/ema/dto/LoginResponseDto.java
@@ -1,9 +1,10 @@
 package com.ema.dto;
 
 import lombok.AllArgsConstructor;
+import lombok.Getter;
 import lombok.Setter;
 
-@Setter
+@Getter
 @AllArgsConstructor
 public class LoginResponseDto {
