+++ src/main/java/com/ema/exception/GlobalExceptionHandler.java
@@ -3,6 +3,7 @@ package com.ema.exception;
 import com.ema.entity.ErrorResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
@@ -15,6 +16,24 @@ public class GlobalExceptionHandler {
         errorResponse.setDebugMessage("Mail not sent");
         errorResponse.setHttpStatus(HttpStatus.FAILED_DEPENDENCY);
         return new ResponseEntity<>(errorResponse, errorResponse.getHttpStatus());
+    }
+
+//    @ExceptionHandler(com.ema.exception.UsernameNotFoundException.class)
+//    public ResponseEntity<ErrorResponse> handlerForUsernameNotFoundException(final UsernameNotFoundException ex){
+//        ErrorResponse errorResponse = new ErrorResponse();
+//        errorResponse.setMessage(ex.getMessage());
+//        errorResponse.setDebugMessage("Username not found, please check your input");
+//        errorResponse.setHttpStatus(HttpStatus.NOT_FOUND);
+//        return new ResponseEntity<>(errorResponse, errorResponse.getHttpStatus());
+//    }
+
+    @ExceptionHandler(VerifyEmailException.class)
+    public ResponseEntity<ErrorResponse> handlerForVerifyEmailException(final VerifyEmailException ex){
+        ErrorResponse errorResponse = new ErrorResponse();
+        errorResponse.setMessage(ex.getMessage());
+        errorResponse.setDebugMessage("Please verify your email");
+        errorResponse.setHttpStatus(HttpStatus.NOT_FOUND);
+        return new ResponseEntity<>(errorResponse, HttpStatus.BAD_REQUEST);
     }
 
     @ExceptionHandler(DepartmentAlreadyExists.class)
