+++ src/main/java/com/ema/exception/VerifyEmailException.java
@@ -1,4 +1,4 @@
-package com.ema.exceptions;
+package com.ema.exception;
 
 public class VerifyEmailException extends RuntimeException{
     public VerifyEmailException(String message) {
