--- src/main/java/com/ema/exceptions/GlobalErrorHandler.java
@@ -1,31 +1,0 @@
-package com.ema.exceptions;
-
-import com.ema.entity.ErrorResponse;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-
-@RestControllerAdvice
-public class GlobalErrorHandler extends ResponseEntityExceptionHandler {
-
-    @ExceptionHandler
-    public ResponseEntity<ErrorResponse> handlerForUsernameNotFoundException(final UsernameNotFoundException ex){
-        ErrorResponse errorResponse = new ErrorResponse();
-        errorResponse.setMessage(ex.getMessage());
-        errorResponse.setDebugMessage("Username not found, please check your input");
-        errorResponse.setHttpStatus(HttpStatus.NOT_FOUND);
-        return new ResponseEntity<>(errorResponse, HttpStatus.BAD_REQUEST);
-    }
-
-    @ExceptionHandler
-    public ResponseEntity<ErrorResponse> handlerForVerifyEmailException(final VerifyEmailException ex){
-        ErrorResponse errorResponse = new ErrorResponse();
-        errorResponse.setMessage(ex.getMessage());
-        errorResponse.setDebugMessage("Please verify your email");
-        errorResponse.setHttpStatus(HttpStatus.NOT_FOUND);
-        return new ResponseEntity<>(errorResponse, HttpStatus.BAD_REQUEST);
-    }
-}
