--- src/main/java/com/ema/exceptions/UsernameNotFoundException.java
@@ -1,8 +1,0 @@
-package com.ema.exceptions;
-
-public class UsernameNotFoundException extends RuntimeException{
-
-    public UsernameNotFoundException(String message) {
-        super(message);
-    }
-}
