+++ src/main/java/com/ema/service/serviceImpl/AuthenticationServiceImpl.java
@@ -3,8 +3,8 @@ package com.ema.service.serviceImpl;
 import com.ema.dto.LoginRequestPayLoad;
 import com.ema.entity.User;
 import com.ema.enums.UserStatus;
-import com.ema.exceptions.UsernameNotFoundException;
-import com.ema.exceptions.VerifyEmailException;
+import com.ema.exception.UserNotFound;
+import com.ema.exception.VerifyEmailException;
 import com.ema.repository.UserRepository;
 import com.ema.security.filter.JwtUtils;
 import com.ema.service.AuthenticationService;
@@ -13,6 +13,8 @@ import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import javax.servlet.http.HttpSession;
@@ -29,12 +31,25 @@ public class AuthenticationServiceImpl implements AuthenticationService {
 
     private final AuthenticationManager authenticationManager;
 
+    private final BCryptPasswordEncoder encoder;
+
     private final HttpSession httpSession;
 
     @Override
     public String login(LoginRequestPayLoad loginRequestPayLoad) {
-        User user = userRepository.findByEmail(loginRequestPayLoad.getEmail())
-                .orElseThrow(() -> new UsernameNotFoundException("user not found!"));
+        User user = userRepository.findByEmail(loginRequestPayLoad.getEmail());
+
+        if (user == null) {
+            throw new UsernameNotFoundException("user not found!");
+        }
+
+        if (!encoder.matches(loginRequestPayLoad.getPassword(), user.getPassword())) {
+            throw new UserNotFound("invalid login credentials!");
+        }
+
+        if (!user.getEmail().equals(loginRequestPayLoad.getEmail())) {
+            throw new UserNotFound("invalid login credentials!");
+        }
 
         if(user.getUserStatus() != UserStatus.ACTIVE){
             throw new VerifyEmailException("Please verify your email!");
